% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBinaryMatrixIO.R
\name{readBinaryIntMatrix}
\alias{readBinaryIntMatrix}
\title{Reads binary matrix}
\usage{
readBinaryIntMatrix(fileStub, nCol, nRow)
}
\arguments{
\item{fileStub}{file name for binary file. Will create fileStub_BY_COL.bin and fileStub_BY_ROW.}

\item{nCol}{number of column}

\item{nRow}{number of rows}
}
\value{
status.  0 =ok
}
\description{
The function checks if the target binary file has the expected number of rows and columns.
}
\examples{

}
