% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBinaryMatrixIO.R
\name{readBinaryIntMatrixByColumn}
\alias{readBinaryIntMatrixByColumn}
\title{Reads binary matrix restricting to the specified columns}
\usage{
readBinaryIntMatrixByColumn(fileStub, nCol, nRow, idxColumn)
}
\arguments{
\item{fileStub}{file name for binary file. Will create fileStub_BY_COL.bin and fileStub_BY_ROW.}

\item{nCol}{number of column}

\item{nRow}{number of rows}

\item{idxColumn}{1 based row indices.}
}
\value{
status.  0 =ok
}
\description{
Reads binary matrix restricting to the specified columns
}
\examples{

}
