% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBinaryMatrixIO.R
\name{readBinaryIntMatrixByRow}
\alias{readBinaryIntMatrixByRow}
\title{Reads binary matrix restricting to the specified rows}
\usage{
readBinaryIntMatrixByRow(fileStub, nCol, nRow, idxRow)
}
\arguments{
\item{fileStub}{file name for binary file. Will create fileStub_BY_COL.bin and fileStub_BY_ROW.}

\item{nCol}{number of column}

\item{nRow}{number of rows}

\item{idxRow}{1 based row indices.}
}
\value{
status.  0 =ok
}
\description{
Reads binary matrix restricting to the specified rows
}
\examples{

}
