% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBinaryMatrixIO.R
\name{writeBinaryIntMatrix}
\alias{writeBinaryIntMatrix}
\title{Writes binary matrix}
\usage{
writeBinaryIntMatrix(fileStub, inmatrix)
}
\arguments{
\item{fileStub}{file name for binary file. Will create fileStub_BY_COL.bin and fileStub_BY_ROW.}

\item{inmatrix}{matrix if integers.  Should have at least one element.}
}
\value{
status.  0 =ok
}
\description{
Writes binary matrix
}
\examples{

}
